requirements.txt
imaplib2==3.6
PyPDF2==3.0.1 
python-docx==0.8.11 
openpyxl==3.1.2 
pandas==2.0.3 
email-validator==2.0.0 
python-dotenv==1.0.0 
nltk==3.8.1 
spacy==3.6.1 
regex==2023.8.8 
beautifulsoup4==4.12.2 
